fixed compatibility with mojolicious +3.21
In Issue #7, kraih writes:

I'm afraid in Mojolicious 3.21 we've had to change the way servers detect applications. Because, as it turns out, the trick we've used before is not actually a Perl feature and will stop working in all versions after 5.17.3. So, from now on you'll have to pass the application instance to the server in your fastcgi command.